refactor: :fire: Change ts-paths
