+++ .vscode/tasks.json
@@ -14,7 +14,7 @@
     {
       "type": "shell",
       "label": "Lint",
-      "command": "pnpm lintfix"
+      "command": "pnpm lint:fix"
     },
     {
       "type": "shell",
@@ -29,7 +29,7 @@
     {
       "type": "shell",
       "label": "CI",
-      "command": "pnpm lintfix && pnpm type:check && pnpm test"
+      "command": "pnpm lint:fix && pnpm type:check && pnpm test"
     }
   ]
 }
