+++ package.json
@@ -9,7 +9,7 @@
     "start:prod": "vite preview",
     "test": "vitest --passWithNoTests",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx --max-warnings=0",
-    "lintfix": "eslint . --ext .js,.jsx,.ts,.tsx --fix --max-warnings=0",
+    "lint:fix": "eslint . --ext .js,.jsx,.ts,.tsx --fix --max-warnings=0",
     "type:check": "tsc --noemit",
     "format": "prettier --write './**/*.{js,jsx,ts,tsx,css,md,json}' --config ./.prettierrc"
   },
