+++ src/App.test.tsx
@@ -1,5 +1,5 @@
-import { App } from 'App';
 import { render } from '@testing-library/react';
+import { App } from './App';
 
 describe('App', () => {
   it('render', () => {
