+++ src/App.tsx
@@ -1,7 +1,7 @@
-import { ThemeProvider } from 'context/ThemeContext/ThemeContext';
+import { LanguageProvider } from '@context/LanguageContext';
+import { ThemeProvider } from '@context/ThemeContext';
+import { Routes } from '@routes';
 import { CookiesProvider } from 'react-cookie';
-import { Routes } from 'routes/Routes';
-import { LanguageProvider } from 'context/LanguageContext/LanguageContext';
 
 export const App = () => {
   return (
