+++ src/context/LanguageContext/index.ts
@@ -1,0 +1,2 @@
+export * from './LanguageContext';
+export * from './language.types';
