+++ src/context/ThemeContext/theme/index.ts
@@ -1,0 +1,3 @@
+export * from './dark-theme';
+export * from './light-theme';
+export * from './theme-options';
