+++ src/context/ThemeContext/theme/light-theme.ts
@@ -1,5 +1,5 @@
-import { components, generalSettings, palette } from 'context/ThemeContext/theme/theme-options';
 import { createTheme } from '@mui/material/styles';
+import { components, generalSettings, palette } from './theme-options';
 
 const scrollbarGrey = {
   100: '#f1f1f1',
